split taskqueries
