+++ src/queries/TicketQueries.tsx
@@ -1,6 +1,6 @@
 import { gql } from '@apollo/client';
 
-const GET_TICKETS = gql`
+const GET_ALL_TICKETS = gql`
   query getAllTickets {
     allTickets {
       _id
@@ -58,61 +58,9 @@ const DELETE_TICKET = gql`
   }
 `;
 
-const GET_PROJECTS = gql`
-  query GetTicketsProjects {
-    getAllProjects {
-      _id
-      name
-    }
-  }
-`;
-
-const GET_PROJECT = gql`
-  query GetOneProject($projectId: String!) {
-    getOneProject(projectId: $projectId) {
-      _id
-      name
-    }
-  }
-`;
-
-const GET_USERS = gql`
-  query AllTicketsUsers {
-    allUsers {
-      _id
-      firstname
-      lastname
-    }
-  }
-`;
-
-const GET_ONE_USER = gql`
-  query GetOneUser($userId: String!) {
-    getOneUser(userId: $userId) {
-      _id
-      firstname
-      lastname
-      email
-      position
-      role
-    }
-  }
-`;
-
-const DELETE_USER = gql`
-  mutation DeleteUser($userId: String!) {
-    deleteUser(UserId: $userId)
-  }
-`;
-
 export {
-  GET_TICKETS,
+  GET_ALL_TICKETS,
   ADD_TICKET,
   UPDATE_TICKET,
   DELETE_TICKET,
-  GET_PROJECTS,
-  GET_PROJECT,
-  GET_USERS,
-  GET_ONE_USER,
-  DELETE_USER,
-};
+}
